gerador_de_dados/requirements.txt
Flask==2.0.3
SQLAlchemy==1.4.25
pyodbc==4.0.32
Werkzeug==2.0.3
PyMySQL==1.0.2
requests
pandas==2.2.2
python-dotenv==1.0.1 
deltalake==0.19.0
pyarrow==17.0.0 